adds gamesessionswitch container split per discussion w ryan
