+++ play/src/App.tsx
@@ -8,11 +8,11 @@ import {
 import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles'; // change to mui v5 see CSS Injection Order section of https://mui.com/material-ui/guides/interoperability/
 import { ApiClient, Environment } from '@righton/networking';
 import PregameContainer from './containers/PregameContainer';
-import { GameInProgressContainer, GameInProgressLoader} from './containers/GameInProgressContainer';
+import GameInProgressContainer from './containers/GameInProgressContainer';
 import Theme from './lib/Theme';
 
 function RedirectToPlayIfMissing() {
-  window.location.href = 'http://central.rightoneducation.com/';
+  window.location.href = 'http://play.rightoneducation.com/';
   return null;
 }
 
@@ -27,8 +27,6 @@ const router = createBrowserRouter(
       <Route 
         path="/game"
         element={<GameInProgressContainer apiClient={apiClient} />}
-        loader={GameInProgressLoader}
-        errorElement={<PregameContainer apiClient={apiClient} isConnectionError/>}
       />
       <Route element={<RedirectToPlayIfMissing />} />
   </>
